Revise and add unit tests in test-pipeline-wrapper.
As a learning exercise as well as to benefit other users of gremlin-node,
this commit implements some of the test documented as unimplemented.

I also revised some existing tests where I felt it was worthwhile
to make the tests produce human readable expected results. This makes
the tests easier to understand by programmers learning gremlin-node.
Additionally, I want the test failures to be more actionable, in anticipation
of the fact that I may in the future do refactorings, and want a better
safety net to ensure those changes do not introduce regressions.