Beginnings of using Q Promises API for some methods.
This is the first commit of several to ehance the gremlin-node API to
support a promises-style API. The plan is that most of the functions that
take callbacks will return promises if a callback is not provided,
but still work as usual if a callback is provided.

Note the use of  the Q nodeify pattern on g.e() and g.v() in this commit.