Change .v() to always return a pipeline (or promise for one).
The .v() step was written to return a vertex if the passed in
arguments included only one id, otherwise returning a pipeline.
This difference in semantics seems troublesome. Code that knows
only one id is being passed in should just call getVertex. Otherwise
the code should expect a pipeline.

This is a change in API so might be controversial, i.e. it might
break existing applications. But given how little activity there
has been with this project in the last ~10 months, I'm betting
it is ok to make API changes like this now.